LPD-72883 move all comments methods to CommentResourceImpl
The methods related to comments in the ObjectEntryResourceImpl class has been moved to the new CommentResourceImpl class

To have a separate JAX-RS resource and be able to implement the CommentEntityModel, following the pattern of the CollaboratorResourceImpl class

----
