+++ modules/apps/object/object-rest-impl/src/main/java/com/liferay/object/rest/internal/resource/v1_0/ObjectEntryResourceImpl.java
@@ -6,10 +6,6 @@
 package com.liferay.object.rest.internal.resource.v1_0;
 
 import com.liferay.exportimport.vulcan.batch.engine.ExportImportVulcanBatchEngineTaskItemDelegate;
-import com.liferay.headless.delivery.dto.v1_0.Comment;
-import com.liferay.headless.delivery.dto.v1_0.Creator;
-import com.liferay.headless.delivery.dto.v1_0.util.CommentUtil;
-import com.liferay.headless.delivery.resource.v1_0.util.CommentResourceUtil;
 import com.liferay.object.constants.ObjectFieldConstants;
 import com.liferay.object.exception.ObjectEntryValidationException;
 import com.liferay.object.model.ObjectDefinition;
@@ -40,32 +36,21 @@ import com.liferay.object.tree.Tree;
 import com.liferay.petra.function.UnsafeFunction;
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringBundler;
-import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.comment.CommentManager;
-import com.liferay.portal.kernel.comment.Discussion;
-import com.liferay.portal.kernel.comment.DiscussionComment;
 import com.liferay.portal.kernel.comment.DiscussionPermission;
-import com.liferay.portal.kernel.exception.NoSuchModelException;
 import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.search.Sort;
 import com.liferay.portal.kernel.search.filter.Filter;
-import com.liferay.portal.kernel.security.auth.PrincipalThreadLocal;
-import com.liferay.portal.kernel.security.permission.ActionKeys;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
 import com.liferay.portal.kernel.security.permission.ResourceActionsUtil;
-import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.HashMapBuilder;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.ParamUtil;
-import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.odata.entity.EntityModel;
 import com.liferay.portal.vulcan.aggregation.Aggregation;
 import com.liferay.portal.vulcan.dto.converter.DTOConverterRegistry;
@@ -77,7 +62,6 @@ import com.liferay.portal.vulcan.resource.NestedFieldsContextResource;
 import com.liferay.portal.vulcan.util.NestedFieldsContextUtil;
 import com.liferay.translation.manager.TranslationManager;
 
-import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.NotSupportedException;
 import jakarta.ws.rs.core.Context;
 import jakarta.ws.rs.core.HttpHeaders;
@@ -92,7 +76,6 @@ import java.util.Collection;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.function.Function;
 
 /**
  * @author Javier Gamarra
@@ -351,33 +334,6 @@ public class ObjectEntryResourceImpl
 	}
 
 	@Override
-	public void deleteByExternalReferenceCodeComment(
-			String externalReferenceCode, String commentExternalReferenceCode)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(externalReferenceCode, null);
-
-		com.liferay.portal.kernel.comment.Comment serviceBuilderComment =
-			_fetchComment(
-				_objectDefinition.getClassName(), objectEntry.getId(),
-				commentExternalReferenceCode,
-				_getNonzeroGroupId(objectEntry.getId()));
-
-		if (serviceBuilderComment == null) {
-			throw new NotFoundException();
-		}
-
-		_deleteComment(serviceBuilderComment.getCommentId());
-	}
-
-	@Override
 	public void deleteObjectEntry(Long objectEntryId) throws Exception {
 		DefaultObjectEntryManager defaultObjectEntryManager =
 			DefaultObjectEntryManagerProvider.provide(
@@ -448,34 +404,6 @@ public class ObjectEntryResourceImpl
 	}
 
 	@Override
-	public void deleteScopeScopeKeyByExternalReferenceCodeComment(
-			String scopeKey, String externalReferenceCode,
-			String commentExternalReferenceCode)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(
-			externalReferenceCode, scopeKey);
-
-		com.liferay.portal.kernel.comment.Comment serviceBuilderComment =
-			_fetchComment(
-				_objectDefinition.getClassName(), objectEntry.getId(),
-				commentExternalReferenceCode, objectEntry.getScopeId());
-
-		if (serviceBuilderComment == null) {
-			throw new NotFoundException();
-		}
-
-		_deleteComment(serviceBuilderComment.getCommentId());
-	}
-
-	@Override
 	public ObjectEntry getApprovedByExternalReferenceCode(
 			String externalReferenceCode)
 		throws Exception {
@@ -541,137 +469,6 @@ public class ObjectEntryResourceImpl
 		return defaultObjectEntryManager.getObjectEntryByVersion(
 			_getDTOConverterContext(null), externalReferenceCode,
 			_objectDefinition, null, version);
-	}
-
-	@Override
-	public Comment getByExternalReferenceCodeComment(
-			String externalReferenceCode, String commentExternalReferenceCode)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(externalReferenceCode, null);
-
-		com.liferay.portal.kernel.comment.Comment serviceBuilderComment =
-			_fetchComment(
-				_objectDefinition.getClassName(), objectEntry.getId(),
-				commentExternalReferenceCode,
-				_getNonzeroGroupId(objectEntry.getId()));
-
-		if (serviceBuilderComment == null) {
-			throw new NoSuchModelException(
-				StringBundler.concat(
-					"No comment exists with object entry ERC ",
-					externalReferenceCode, " and ERC ",
-					commentExternalReferenceCode));
-		}
-
-		_discussionPermission.checkViewPermission(
-			PermissionThreadLocal.getPermissionChecker(),
-			contextCompany.getCompanyId(), serviceBuilderComment.getGroupId(),
-			serviceBuilderComment.getClassName(),
-			serviceBuilderComment.getClassPK());
-
-		return CommentUtil.toComment(
-			serviceBuilderComment, _commentManager, PortalUtil.getPortal());
-	}
-
-	@Override
-	public Page<Comment> getByExternalReferenceCodeCommentChildCommentsPage(
-			String externalReferenceCode, String commentExternalReferenceCode,
-			String search, Aggregation aggregation, Filter filter,
-			Pagination pagination, Sort[] sorts)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(externalReferenceCode, null);
-
-		com.liferay.portal.kernel.comment.Comment serviceBuilderComment =
-			_fetchComment(
-				_objectDefinition.getClassName(), objectEntry.getId(),
-				commentExternalReferenceCode,
-				_getNonzeroGroupId(objectEntry.getId()));
-
-		if (serviceBuilderComment == null) {
-			throw new NotFoundException();
-		}
-
-		return CommentResourceUtil.getComments(
-			HashMapBuilder.put(
-				"create",
-				addAction(
-					ActionKeys.ADD_DISCUSSION,
-					"postByExternalReferenceCodeCommentChildComment",
-					_objectDefinition.getClassName(), null)
-			).put(
-				"delete",
-				addAction(
-					ActionKeys.DELETE, "deleteByExternalReferenceCodeComment",
-					_objectDefinition.getClassName(), null)
-			).build(),
-			serviceBuilderComment.getCommentId(), contextCompany.getCompanyId(),
-			_commentManager, search, aggregation, filter, pagination,
-			PortalUtil.getPortal(), sorts);
-	}
-
-	@Override
-	public Page<Comment> getByExternalReferenceCodeCommentsPage(
-			String externalReferenceCode, String search,
-			Aggregation aggregation, Filter filter, Pagination pagination,
-			Sort[] sorts)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(externalReferenceCode, null);
-
-		long groupId = _getNonzeroGroupId(objectEntry.getId());
-
-		long objectEntryId = objectEntry.getId();
-
-		Discussion discussion = _commentManager.getDiscussion(
-			PrincipalThreadLocal.getUserId(), groupId,
-			_objectDefinition.getClassName(), objectEntry.getId(),
-			_createServiceContextFunction());
-
-		DiscussionComment rootDiscussionComment =
-			discussion.getRootDiscussionComment();
-
-		Creator creator = objectEntry.getCreator();
-
-		return CommentResourceUtil.getComments(
-			HashMapBuilder.put(
-				"add-discussion",
-				addAction(
-					ActionKeys.ADD_DISCUSSION, objectEntryId,
-					"postByExternalReferenceCodeComment", creator.getId(),
-					_objectDefinition.getClassName(), groupId)
-			).put(
-				"get",
-				addAction(
-					ActionKeys.VIEW, objectEntryId,
-					"getByExternalReferenceCodeCommentsPage", creator.getId(),
-					_objectDefinition.getClassName(), groupId)
-			).build(),
-			rootDiscussionComment.getCommentId(), contextCompany.getCompanyId(),
-			_commentManager, search, aggregation, filter, pagination,
-			PortalUtil.getPortal(), sorts);
 	}
 
 	@Override
@@ -964,92 +761,6 @@ public class ObjectEntryResourceImpl
 	}
 
 	@Override
-	public Comment getScopeScopeKeyByExternalReferenceCodeComment(
-			String scopeKey, String externalReferenceCode,
-			String commentExternalReferenceCode)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(
-			externalReferenceCode, scopeKey);
-
-		com.liferay.portal.kernel.comment.Comment serviceBuilderComment =
-			_fetchComment(
-				_objectDefinition.getClassName(), objectEntry.getId(),
-				commentExternalReferenceCode, objectEntry.getScopeId());
-
-		if (serviceBuilderComment == null) {
-			throw new NoSuchModelException(
-				StringBundler.concat(
-					"No comment exists with scopeKey ", scopeKey,
-					", object entry ERC ", externalReferenceCode, " and ERC ",
-					commentExternalReferenceCode));
-		}
-
-		_discussionPermission.checkViewPermission(
-			PermissionThreadLocal.getPermissionChecker(),
-			contextCompany.getCompanyId(), serviceBuilderComment.getGroupId(),
-			serviceBuilderComment.getClassName(),
-			serviceBuilderComment.getClassPK());
-
-		return CommentUtil.toComment(
-			serviceBuilderComment, _commentManager, PortalUtil.getPortal());
-	}
-
-	@Override
-	public Page<Comment>
-			getScopeScopeKeyByExternalReferenceCodeCommentChildCommentsPage(
-				String scopeKey, String externalReferenceCode,
-				String commentExternalReferenceCode, String search,
-				Aggregation aggregation, Filter filter, Pagination pagination,
-				Sort[] sorts)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(
-			externalReferenceCode, scopeKey);
-
-		com.liferay.portal.kernel.comment.Comment serviceBuilderComment =
-			_fetchComment(
-				_objectDefinition.getClassName(), objectEntry.getId(),
-				commentExternalReferenceCode, objectEntry.getScopeId());
-
-		if (serviceBuilderComment == null) {
-			throw new NotFoundException();
-		}
-
-		return CommentResourceUtil.getComments(
-			HashMapBuilder.put(
-				"create",
-				addAction(
-					ActionKeys.ADD_DISCUSSION,
-					"postScopeScopeKeyByExternalReferenceCodeComment" +
-						"ChildComment",
-					_objectDefinition.getClassName(), null)
-			).put(
-				"delete",
-				addAction(
-					ActionKeys.DELETE, "deleteByExternalReferenceCodeComment",
-					_objectDefinition.getClassName(), null)
-			).build(),
-			serviceBuilderComment.getCommentId(), contextCompany.getCompanyId(),
-			_commentManager, search, aggregation, filter, pagination,
-			PortalUtil.getPortal(), sorts);
-	}
-
-	@Override
 	public Response getScopeScopeKeyByExternalReferenceCodeTranslation(
 			String scopeKey, String externalReferenceCode,
 			String sourceLanguageId, String targetLanguageIds, String version)
@@ -1199,58 +910,6 @@ public class ObjectEntryResourceImpl
 		return defaultObjectEntryManager.expireObjectEntryByVersion(
 			_getDTOConverterContext(null), externalReferenceCode,
 			_objectDefinition, null, version);
-	}
-
-	@Override
-	public Comment postByExternalReferenceCodeComment(
-			String externalReferenceCode, Comment comment)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(externalReferenceCode, null);
-
-		return _addComment(
-			comment.getExternalReferenceCode(),
-			_getNonzeroGroupId(objectEntry.getId()), null, objectEntry.getId(),
-			comment.getText());
-	}
-
-	@Override
-	public Comment postByExternalReferenceCodeCommentChildComment(
-			String externalReferenceCode, String commentExternalReferenceCode,
-			Comment comment)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(externalReferenceCode, null);
-
-		com.liferay.portal.kernel.comment.Comment serviceBuilderComment =
-			_fetchComment(
-				_objectDefinition.getClassName(), objectEntry.getId(),
-				commentExternalReferenceCode,
-				_getNonzeroGroupId(objectEntry.getId()));
-
-		if (serviceBuilderComment == null) {
-			throw new NotFoundException();
-		}
-
-		return _addComment(
-			comment.getExternalReferenceCode(),
-			serviceBuilderComment.getGroupId(),
-			serviceBuilderComment.getCommentId(), objectEntry.getId(),
-			comment.getText());
 	}
 
 	@Override
@@ -1494,58 +1153,6 @@ public class ObjectEntryResourceImpl
 	}
 
 	@Override
-	public Comment postScopeScopeKeyByExternalReferenceCodeComment(
-			String scopeKey, String externalReferenceCode, Comment comment)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(
-			externalReferenceCode, scopeKey);
-
-		return _addComment(
-			comment.getExternalReferenceCode(), objectEntry.getScopeId(), null,
-			objectEntry.getId(), comment.getText());
-	}
-
-	@Override
-	public Comment postScopeScopeKeyByExternalReferenceCodeCommentChildComment(
-			String scopeKey, String externalReferenceCode,
-			String commentExternalReferenceCode, Comment comment)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(
-			externalReferenceCode, scopeKey);
-
-		com.liferay.portal.kernel.comment.Comment serviceBuilderComment =
-			_fetchComment(
-				_objectDefinition.getClassName(), objectEntry.getId(),
-				commentExternalReferenceCode, objectEntry.getScopeId());
-
-		if (serviceBuilderComment == null) {
-			throw new NotFoundException();
-		}
-
-		return _addComment(
-			comment.getExternalReferenceCode(),
-			serviceBuilderComment.getGroupId(),
-			serviceBuilderComment.getCommentId(), objectEntry.getId(),
-			comment.getText());
-	}
-
-	@Override
 	public ObjectEntry postScopeScopeKeyByExternalReferenceCodeExpire(
 			String scopeKey, String externalReferenceCode)
 		throws Exception {
@@ -1643,37 +1250,6 @@ public class ObjectEntryResourceImpl
 		return defaultObjectEntryManager.restoreObjectEntryByVersion(
 			_getDTOConverterContext(null), externalReferenceCode,
 			_objectDefinition, null, version);
-	}
-
-	@Override
-	public Comment putByExternalReferenceCodeComment(
-			String externalReferenceCode, String commentExternalReferenceCode,
-			Comment comment)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(externalReferenceCode, null);
-
-		long groupId = _getNonzeroGroupId(objectEntry.getId());
-
-		com.liferay.portal.kernel.comment.Comment serviceBuilderComment =
-			_fetchComment(
-				_objectDefinition.getClassName(), objectEntry.getId(),
-				commentExternalReferenceCode, groupId);
-
-		if (serviceBuilderComment != null) {
-			return _updateComment(comment, serviceBuilderComment);
-		}
-
-		return _addComment(
-			commentExternalReferenceCode, groupId, null, objectEntry.getId(),
-			comment.getText());
 	}
 
 	@Override
@@ -1809,36 +1385,6 @@ public class ObjectEntryResourceImpl
 	}
 
 	@Override
-	public Comment putScopeScopeKeyByExternalReferenceCodeComment(
-			String scopeKey, String externalReferenceCode,
-			String commentExternalReferenceCode, Comment comment)
-		throws Exception {
-
-		if (!_objectDefinition.isEnableComments() ||
-			!FeatureFlagManagerUtil.isEnabled(
-				_objectDefinition.getCompanyId(), "LPD-69419")) {
-
-			throw new UnsupportedOperationException();
-		}
-
-		ObjectEntry objectEntry = _getObjectEntry(
-			externalReferenceCode, scopeKey);
-
-		com.liferay.portal.kernel.comment.Comment serviceBuilderComment =
-			_fetchComment(
-				_objectDefinition.getClassName(), objectEntry.getId(),
-				commentExternalReferenceCode, objectEntry.getScopeId());
-
-		if (serviceBuilderComment != null) {
-			return _updateComment(comment, serviceBuilderComment);
-		}
-
-		return _addComment(
-			commentExternalReferenceCode, objectEntry.getScopeId(), null,
-			objectEntry.getId(), comment.getText());
-	}
-
-	@Override
 	public void
 			putScopeScopeKeyByExternalReferenceCodeObjectActionObjectActionName(
 				String scopeKey, String externalReferenceCode,
@@ -1939,79 +1485,12 @@ public class ObjectEntryResourceImpl
 		}
 	}
 
-	private Comment _addComment(
-			String externalReferenceCode, long groupId, Long parentCommentId,
-			long objectEntryId, String text)
-		throws Exception {
-
-		_discussionPermission.checkAddPermission(
-			PermissionThreadLocal.getPermissionChecker(),
-			contextCompany.getCompanyId(), groupId,
-			_objectDefinition.getClassName(), objectEntryId);
-
-		if (parentCommentId != null) {
-			return CommentUtil.toComment(
-				() -> _commentManager.fetchComment(
-					_commentManager.addComment(
-						externalReferenceCode, PrincipalThreadLocal.getUserId(),
-						_objectDefinition.getClassName(), objectEntryId,
-						StringPool.BLANK, parentCommentId, StringPool.BLANK,
-						StringBundler.concat("<p>", text, "</p>"),
-						_createServiceContextFunction())),
-				_commentManager, PortalUtil.getPortal());
-		}
-
-		return CommentUtil.toComment(
-			() -> _commentManager.fetchComment(
-				_commentManager.addComment(
-					externalReferenceCode, PrincipalThreadLocal.getUserId(),
-					groupId, _objectDefinition.getClassName(), objectEntryId,
-					StringPool.BLANK, StringPool.BLANK,
-					StringBundler.concat("<p>", text, "</p>"),
-					_createServiceContextFunction())),
-			_commentManager, PortalUtil.getPortal());
-	}
-
 	private void _checkFeatureFlag() {
 		if (!FeatureFlagManagerUtil.isEnabled(
 				_objectDefinition.getCompanyId(), "LPD-17564")) {
 
 			throw new UnsupportedOperationException();
 		}
-	}
-
-	private Function<String, ServiceContext> _createServiceContextFunction() {
-		return className -> {
-			ServiceContext serviceContext = new ServiceContext();
-
-			serviceContext.setWorkflowAction(WorkflowConstants.ACTION_PUBLISH);
-
-			return serviceContext;
-		};
-	}
-
-	private void _deleteComment(Long commentId) throws Exception {
-		_discussionPermission.checkDeletePermission(
-			PermissionThreadLocal.getPermissionChecker(), commentId);
-
-		_commentManager.deleteComment(commentId);
-	}
-
-	private com.liferay.portal.kernel.comment.Comment _fetchComment(
-		String className, long classPK, String externalReferenceCode,
-		long groupId) {
-
-		com.liferay.portal.kernel.comment.Comment serviceBuilderComment =
-			_commentManager.fetchComment(groupId, externalReferenceCode);
-
-		if ((serviceBuilderComment != null) &&
-			CommentResourceUtil.isAssociated(
-				className, classPK, serviceBuilderComment)) {
-
-			return serviceBuilderComment;
-		}
-
-		return null;
 	}
 
 	private DefaultDTOConverterContext _getDTOConverterContext(
@@ -2081,28 +1560,6 @@ public class ObjectEntryResourceImpl
 			"(", filterString, ") and ", assigneeFilterString);
 	}
 
-	private long _getNonzeroGroupId(long objectEntryId) throws Exception {
-		com.liferay.object.model.ObjectEntry objectEntry =
-			_objectEntryLocalService.getObjectEntry(objectEntryId);
-
-		return objectEntry.getNonzeroGroupId();
-	}
-
-	private ObjectEntry _getObjectEntry(
-			String objectEntryExternalReferenceCode, String scopeKey)
-		throws Exception {
-
-		DefaultObjectEntryManager defaultObjectEntryManager =
-			DefaultObjectEntryManagerProvider.provide(
-				_objectEntryManagerRegistry.getObjectEntryManager(
-					_objectDefinition.getCompanyId(),
-					_objectDefinition.getStorageType()));
-
-		return defaultObjectEntryManager.getObjectEntry(
-			contextCompany.getCompanyId(), _getDTOConverterContext(null),
-			objectEntryExternalReferenceCode, _objectDefinition, scopeKey);
-	}
-
 	private String _getScopeKey(Map<String, Serializable> parameters) {
 		if (parameters.containsKey("scopeKey")) {
 			return String.valueOf(parameters.get("scopeKey"));
@@ -2131,27 +1588,6 @@ public class ObjectEntryResourceImpl
 		}
 
 		return null;
-	}
-
-	private Comment _updateComment(
-			Comment comment,
-			com.liferay.portal.kernel.comment.Comment serviceBuilderComment)
-		throws Exception {
-
-		_discussionPermission.checkUpdatePermission(
-			PermissionThreadLocal.getPermissionChecker(),
-			serviceBuilderComment.getCommentId());
-
-		return CommentUtil.toComment(
-			() -> _commentManager.fetchComment(
-				_commentManager.updateComment(
-					PrincipalThreadLocal.getUserId(),
-					_objectDefinition.getClassName(),
-					serviceBuilderComment.getClassPK(),
-					serviceBuilderComment.getCommentId(), StringPool.BLANK,
-					StringBundler.concat("<p>", comment.getText(), "</p>"),
-					_createServiceContextFunction())),
-			_commentManager, PortalUtil.getPortal());
 	}
 
 	private ValidationResponse _validateObjectEntry(
